fix: clippy clean under Rust 1.95
The toolchain moved and CI installs `stable` unpinned, so `run-tests.sh`
and the CI clippy leg went red on lints that did not exist when the code
was written: 26 `needless_borrow`, 13 `format!`-with-trailing-newline,
plus `manual_contains`, `redundant_closure`, `useless_conversion` and
friends. All mechanical, applied via `cargo clippy --fix`.

Four sites are deliberate rather than mechanical:

* Three error enums (`BootError`, `FromArchiveBytesError`,
  `ComposeHealthError`) carry `#[allow(clippy::large_enum_variant)]`.
  Their width comes from lifting lower-layer error types verbatim
  through `#[from]` — the property that makes the single-envelope
  design work. Boxing would trade that for a byte count on a path
  that runs at most once per call.
* `gather_from_store` carries `#[allow(clippy::too_many_arguments)]`:
  five of its ten parameters are the engine-shaped callbacks that keep
  it engine-agnostic, and a struct for one call site moves the arity
  rather than removing it.
* `MountedBackend` becomes `pub(crate)` so the crate-internal
  `build_mem_router_from_mounts` may name it in its signature. Not
  re-exported — the type stays invisible outside the crate.

`clippy --fix` also stripped the `RelateOpInput` import from both MCP
servers, where it is used only under `#[cfg(test)]`; the import moved
into each test module instead of being deleted.

+++ crates/memstead-base/src/engine/due.rs
@@ -71,7 +71,7 @@ fn parse_ymd(s: &str) -> Option<(i64, u32, u32)> {
 fn days_from_civil(y: i64, m: u32, d: u32) -> i64 {
     let y = if m <= 2 { y - 1 } else { y };
     let era = if y >= 0 { y } else { y - 399 } / 400;
-    let yoe = (y - era * 400) as i64;
+    let yoe = y - era * 400;
     let mp = ((m + 9) % 12) as i64;
     let doy = (153 * mp + 2) / 5 + d as i64 - 1;
     let doe = yoe * 365 + yoe / 4 - yoe / 100 + doy;
